requirements-airflow.txt
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-cncf-kubernetes>=7.0.0
apache-airflow-providers-docker>=3.0.0
kubernetes>=28.0.0
psycopg2-binary==2.9.9
pandas==1.5.3
numpy==1.24.3
lightgbm==4.0.0
scikit-learn==1.3.2
mlflow==2.9.1
pyyaml==6.0

